Call notification not displayed at SIM query UI
When the SIM slection query UI is launched from InCallActivty and
when the device receives incoming call, the current state of call ui
considered as IN_CALL, due to this notification is not displaying.
Audio is not heard due to the mismatch in the foreground call. check
whether SIM selection ui is diplayed and when the call state is
SELECT_PHONE_ACCOUNT return false for isAlive query.

CRs-Fixed: 1014066
Change-Id: I09ba2bae4d3348d774c78311ef96fc396f85823f