IMS-VT: Fix orientation change issue during VT conference call.
Current orientation in OrientationModeHandler is dependent only
on the cached orientationMode value. Where as orientation value for
a call depends on the orientation mode as well as videostate.
So if the primary call changes from audio only call to video call
while orientation mode remains as dynamic, VideoCallPresenter is not
setting the new orientation.
To fix this store that last set orientation value in
VideoCallPresenter and use it for checking orientation mode change.

Change-Id: Ifff35049ae3f640275371cd1d17ee57f83cc4600
CRs-Fixed: 1047384